obs-ffmpeg: add an autoplay flag for ffmpeg media
When the autoplay flag is true, the media will start playing
immediately upon load.

When the autoplay flag is false, the media needs to be explicitly
triggered (e.g. by the user pressing play).

For backwards compatibility, this flag is ignored when "restart
playback when source becomes active" is enabled.